@@ -13,9 +13,7 @@
     "@inertiajs/inertia": "^0.1.7",
     "@inertiajs/inertia-vue": "^0.1.2",
     "axios": "^0.19",
-    "bootstrap": "^4.0.0",
     "cross-env": "^7.0",
-    "jquery": "^3.2",
     "laravel-mix": "^5.0.1",
     "laravel-mix-purgecss": "^4.2.0",
     "lodash": "^4.17.13",
@@ -23,12 +21,11 @@
     "postcss-import": "^12.0.1",
     "postcss-nesting": "^7.0.0",
     "resolve-url-loader": "^2.3.1",
-    "sass": "^1.20.1",
-    "sass-loader": "^8.0.0",
-    "vue": "^2.5.17",
-    "vue-template-compiler": "^2.6.10"
+    "vue": "^2.6.11",
+    "vue-template-compiler": "^2.6.11"
   },
   "dependencies": {
-    "tailwindcss": "^1.3.4"
+    "@popperjs/core": "^2.4.0",
+    "tailwindcss": "^1.4.6"
   }
 }
