@@ -6,17 +6,7 @@
 
 require('./bootstrap')
 
-window.Vue = require('vue')
-
-import './OutsideClick'
-
-/**
- * The following block of code may be used to automatically register your
- * Vue components. It will recursively scan this directory for the Vue
- * components and automatically register them with their "basename".
- *
- * Eg. ./components/ExampleComponent.vue -> <example-component></example-component>
- */
+import Vue from 'vue'
 
 const files = require.context('./', true, /\.vue$/i)
 files.keys().map(key =>
@@ -28,14 +18,6 @@ files.keys().map(key =>
     files(key).default
   )
 )
-
-// Vue.component('example-component', require('./components/ExampleComponent.vue').default);
-
-/**
- * Next, we will create a fresh Vue application instance and attach it to
- * the page. Then, you may begin adding components to this application
- * or customize the JavaScript scaffolding to fit your unique needs.
- */
 
 const app = new Vue({
   el: '#app',
