@@ -1,5 +1,15 @@
 module.exports = {
-    theme: {},
-    variants: {},
-    plugins: []
-};
+  theme: {},
+  variants: {},
+  plugins: [],
+  purge: {
+    content: [
+      './resources/**/*.js',
+      './resources/**/*.php',
+      './resources/**/*.vue',
+    ],
+    options: {
+      whitelistPatterns: [/nprogress/],
+    },
+  },
+}
