@@ -3,7 +3,6 @@
 namespace Tests\Feature;
 
 use Tests\TestCase;
-use Illuminate\Support\Facades\DB;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 
 class ExampleTest extends TestCase
@@ -12,7 +11,6 @@ class ExampleTest extends TestCase
 
     public function testBasicTest()
     {
-        // dump(DB::connection()->getSchemaBuilder()->getColumnListing('users'));
-        // dump(DB::connection()->table('users'));
+        $this->post('/', ['foo' => 'bar']);
     }
 }
