@@ -1,21 +1,12 @@
-const mix = require('laravel-mix')
-const tailwindcss = require('tailwindcss')
-
-/*
- |--------------------------------------------------------------------------
- | Mix Asset Management
- |--------------------------------------------------------------------------
- |
- | Mix provides a clean, fluent API for defining some Webpack build steps
- | for your Laravel application. By default, we are compiling the Sass
- | file for the application as well as bundling up all the JS files.
- |
- */
+let cssImport = require('postcss-import')
+let cssNesting = require('postcss-nesting')
+let mix = require('laravel-mix')
+let tailwindcss = require('tailwindcss')
 
 mix
   .js('resources/js/app.js', 'public/js')
-  .sass('resources/sass/app.scss', 'public/css')
-  .options({
-    processCssUrls: false,
-    postCss: [tailwindcss('./tailwind.config.js')],
-  })
+  .postCss('resources/css/app.css', 'public/css', [
+    cssImport(),
+    cssNesting(),
+    tailwindcss('tailwind.config.js'),
+  ])
